refactor: modernize Go runtime packages (#293)
* refactor: modernize Go runtime

* fix: adjustments

* fix: fence canceled input promotion replays

* fix: few adjustments

* fix: preserve opaque delivery ids

* fix: preserve session and bridge identity contracts

---------

+++ internal/daemon/loop_goal_prompt_recovery.go
@@ -131,7 +131,7 @@ func managedGoalRecoveryOwnerMatches(
 	identity goalpkg.PromptRecoveryIdentity,
 ) bool {
 	return owner.LoopRunID == string(identity.Key.LoopRunID) &&
-		owner.RunGeneration == identity.Key.Generation &&
+		owner.RunGeneration == int64(identity.Key.Generation) &&
 		owner.ControlEpoch == identity.ExpectedControlEpoch &&
 		owner.BindingEpoch == identity.ExpectedBindingEpoch &&
 		owner.PromptID == strings.TrimSpace(identity.PromptID) &&
